pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paper-expert"
version = "0.1.0"
description = "Universal AI-powered academic research assistant with user-defined domain configuration"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "paper-qa>=5",
    "typer[all]>=0.12",
    "rich>=13",
    "httpx>=0.27",
    "bibtexparser>=1.4",
    "pyalex>=0.14",
    "semanticscholar>=0.8",
    "networkx>=3",
    "pyyaml>=6",
    "pydantic>=2",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[project.scripts]
paper-expert = "paper_expert.cli:app"
paper-expert-mcp = "paper_expert.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["paper_expert"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]